requirements.txt
streamlit>=1.22
python-dotenv>=0.19
openai>=1.0
spotipy>=2.23